2012/01/18 v4.31 Fixed a problem whereby sync would fail if there was not .fish/alias tag defined.   It now defines it in this case. (This also caused a problem on starting fish interactively, since that does a sync.)
Moved some code from fish.py to cli.py to facilitate the above change.

Fixed a problem whereby Fluidinfo.tag_by_query would fail under
python2.7 if there were non-ascii characters in tag names or values.
This was because I was lazily constructing a json-like string by hand
instead of using json.dumps.   Oddly, this worked fine in python 2.5, 2.6.

Cleaned up the import structure a bit.

In principle, I would much rather use the requests library rather than
httplib2 in all circumstances, but there are two problems:

  1. It is not universally available (e.g., as far as I can tell, it's
     not available on Google App Engine

  2. Either there's something wrong with 1.0.0 and above or Fish is
     using it wrongly because requests is having a problem iterating
     over the headers Fish is sending it.

For the moment, therefore, I have made it so that tries to import
requests and checks that it is a version under 1.0.0.   If the
library is not present, or is at 1.0.0 or above, Fish falls back to
httplib2.

+++ fishlib.py
@@ -6,10 +6,9 @@
 #               in the AUTHOR
 # Licence terms in LICENCE.
 
-__version__ = u'4.30'
+__version__ = u'4.31'
 VERSION = __version__
 
-import base64
 import codecs
 import cPickle as pickle
 import os
@@ -18,8 +17,14 @@ import sys
 import types
 import urllib
 from functools import wraps
-#from httplib2 import Http
-import requests
+try:
+    import requests
+    if requests.__version__ >= '1.0.0':
+        raise ImportError
+    REQUESTS = True
+except:
+    from httplib2 import Http
+    REQUESTS = False
 import cline
 from cache import Cache
 from fishbase import (get_credentials_file, O, formatted_tag_value,
@@ -437,9 +442,17 @@ class Fluidinfo:
         self.timeout = choose_http_timeout()
         if not host.startswith(u'http'):
             self.host = u'http://%s' % host
-        self.headers = {
-        }
-        self.auth = (credentials.username, credentials.password)
+        if REQUESTS:
+            self.headers = {}
+            self.auth = (credentials.username, credentials.password)
+        else:
+            userpass = u'%s:%s' % (credentials.username, credentials.password)
+            encoded = unicode(userpass.encode('UTF-8').encode('base64').strip())
+            auth = u'Basic %s' % encoded
+            self.headers = {
+                u'Authorization': auth
+            }
+
 
     def Print(self, s, allowSave=True, allowPrint=True):
         if self.saveOutput:
@@ -514,28 +527,31 @@ class Fluidinfo:
         try:
             response, content, result, status = self.request(url, method,
                                                          body8, headers)
-        except:
-            raise Exception(u'URL Fetch failure (URL len %d): %s'
-                            % (len(url), url))
+        except Exception, e:
+            raise Exception(u'URL Fetch failure (URL len %d): %s\n%s'
+                            % (len(url), url, str(e)))
             
         return status, result
 
     def request(self, url, method, body8, headers):
-#        http = _get_http(self.timeout)
-#        response, content = http.request(url, method, body8, headers)
-        if method == 'GET':
-            r = requests.get(url, auth=self.auth, headers=headers)
-        elif method == 'PUT':
-            r = requests.put(url, body8, auth=self.auth, headers=headers)
-        elif method == 'POST':
-            r = requests.post(url, body8, auth=self.auth, headers=headers)
-        elif method == 'DELETE':
-            r = requests.delete(url, auth=self.auth, headers=headers)
+        if REQUESTS:
+            if method == 'GET':
+                r = requests.get(url, auth=self.auth, headers=headers)
+            elif method == 'PUT':
+                r = requests.put(url, body8, auth=self.auth, headers=headers)
+            elif method == 'POST':
+                r = requests.post(url, body8, auth=self.auth, headers=headers)
+            elif method == 'DELETE':
+                r = requests.delete(url, auth=self.auth, headers=headers)
+            else:
+                raise GeneralError('Unknown method')
+            response = r.headers
+            content = r.content
+            status = r.status_code
         else:
-            raise GeneralError('Unknown method')
-        response = r.headers
-        content = r.content
-        status = r.status_code
+            http = _get_http(self.timeout)
+            response, content = http.request(url, method, body8, headers)
+            status = response.status
 
         if response[u'content-type'].startswith(u'application/json'):
             result = json.loads(content)
@@ -1093,10 +1109,11 @@ class Fluidinfo:
 
 
         """
-        strHash = u'{%s}' % u', '.join(u'"%s": {"value": %s}'
-                                       % (tag, format_val(tagsToSet[tag]))
-                                       for tag in tagsToSet)
-        (v, r) = self.call(u'PUT', u'/values', strHash, {u'query': query})
+        h = {}
+        for tag in tagsToSet:
+            h[tag] = {u'value': tagsToSet[tag]}
+        (v, r) = self.call(u'PUT', u'/values', json.dumps(h),
+                          {u'query': query})
         assert_status(v, STATUS.NO_CONTENT)
